Linked to Filter Lambda function + Add new skills
Filtering is now linked to the lambda function. Added an accuracy slider to select how accurate the matching is. When searching for skills to filter by or to add to a new position template, if no results are found, a button displays which adds that skill to the database and selects it.